New invite logic.

+++ app/src/models/invitation.rs
@@ -1,6 +1,6 @@
 use crate::authentication::Authentication;
 use crate::errors::CustomError;
-use crate::models::organisation;
+use crate::models::{organisation, user};
 use rand::Rng;
 use sha2::{Digest, Sha256};
 use sqlx::PgPool;
@@ -9,6 +9,8 @@ pub struct Invitation {
     pub id: i32,
     pub organisation_id: i32,
     pub email: String,
+    pub invitation_selector: String,
+    pub invitation_verifier_hash: String,
     pub created_at: chrono::DateTime<chrono::Utc>,
     pub updated_at: chrono::DateTime<chrono::Utc>,
 }
@@ -22,11 +24,14 @@ impl Invitation {
         let org = organisation::Organisation::get_primary_org(pool, authenticated_user).await?;
 
         let invitation_selector = rand::thread_rng().gen::<[u8; 8]>();
-        let invitation_selector_base64 = base64::encode(invitation_selector);
+        let invitation_selector_base64 =
+            base64::encode_config(invitation_selector, base64::URL_SAFE);
         let invitation_verifier = rand::thread_rng().gen::<[u8; 24]>();
         let invitation_verifier_hash = Sha256::digest(&invitation_verifier);
-        let invitation_verifier_hash_base64 = base64::encode(invitation_verifier_hash);
-        let invitation_verifier_base64 = base64::encode(invitation_verifier);
+        let invitation_verifier_hash_base64 =
+            base64::encode_config(invitation_verifier_hash, base64::URL_SAFE);
+        let invitation_verifier_base64 =
+            base64::encode_config(invitation_verifier, base64::URL_SAFE);
 
         sqlx::query!(
             "
@@ -45,25 +50,66 @@ impl Invitation {
         Ok((invitation_verifier_base64, invitation_selector_base64))
     }
 
-    pub async fn delete_dangerous(
+    pub async fn accept_invitation(
         pool: &PgPool,
-        email: &str,
-        organisation_id: u32,
+        current_user: &Authentication,
+        invitation_selector: &str,
+        invitation_verifier: &str,
     ) -> Result<(), CustomError> {
-        sqlx::query!(
-            r#"
-                DELETE FROM
-                    invitations
+        let invitation_verifier = base64::decode_config(invitation_verifier, base64::URL_SAFE)
+            .map_err(|e| CustomError::FaultySetup(e.to_string()))?;
+        let invitation_verifier_hash = Sha256::digest(&invitation_verifier);
+        let invitation_verifier_hash_base64 =
+            base64::encode_config(invitation_verifier_hash, base64::URL_SAFE);
+
+        let invitation = sqlx::query_as!(
+            Invitation,
+            "
+                SELECT 
+                    id, 
+                    organisation_id, 
+                    email, 
+                    invitation_selector, 
+                    invitation_verifier_hash,
+                    created_at,
+                    updated_at
+                FROM 
+                    invitations 
                 WHERE
-                    email = $1
-                AND
-                    organisation_id = $2
-            "#,
-            email,
-            organisation_id as i32
+                    invitation_selector = $1
+            ",
+            invitation_selector,
         )
-        .execute(pool)
+        .fetch_one(pool)
         .await?;
+
+        if invitation.invitation_verifier_hash == invitation_verifier_hash_base64 {
+            let user = user::User::get_dangerous(pool, current_user.user_id).await?;
+
+            if user.email == invitation.email {
+                organisation::Organisation::add_user_dangerous(
+                    pool,
+                    &invitation.email,
+                    invitation.organisation_id as u32,
+                )
+                .await?;
+
+                sqlx::query!(
+                    r#"
+                        DELETE FROM
+                            invitations
+                        WHERE
+                            email = $1
+                        AND
+                            organisation_id = $2
+                    "#,
+                    &invitation.email,
+                    invitation.organisation_id
+                )
+                .execute(pool)
+                .await?;
+            }
+        }
 
         Ok(())
     }
@@ -80,6 +126,8 @@ impl Invitation {
                 SELECT  
                     id, 
                     email,
+                    invitation_selector, 
+                    invitation_verifier_hash,
                     organisation_id,
                     updated_at, 
                     created_at
